requirements.txt
langchain
langgraph
langchain-google-genai
pandas
python-dotenv
camelot-py[cv] 